+++ bigraph_schema/type_functions.py
@@ -82,7 +82,7 @@ from bigraph_schema.registry import (
     is_schema_key, non_schema_keys, type_parameter_key,
     deep_merge, hierarchy_depth,
     establish_path)
-
+from bigraph_schema.utilities import is_empty, union_keys, tuple_from_type, array_shape, read_datatype, read_shape
 
 # Create a new module dynamically for the dataclasses
 module_name = 'bigraph_schema.data'
@@ -116,12 +116,6 @@ merge_schema_keys = (
     '_ports',
     '_type_parameters',
 )
-
-DTYPE_MAP = {
-    'float': 'float64',
-    'integer': 'int64',
-    'string': 'str'}
-
 
 def diff(a, b):
     if isinstance(a, dict) and isinstance(b, dict):
@@ -180,11 +174,11 @@ def type_merge(dct, merge_dct, path=tuple(), merge_supers=False):
 def get_path(tree, path):
     """
     Given a tree and a path, find the subtree at that path
-    
+
     Args:
     - tree: the tree we are looking in (a nested dict)
     - path: a list/tuple of keys we follow down the tree to find the subtree we are looking for
-    
+
     Returns:
     - subtree: the subtree found by following the list of keys down the tree
     """
@@ -2454,70 +2448,6 @@ def register_base_reactions(core):
     core.register_reaction('remove', remove_reaction)
     core.register_reaction('replace', replace_reaction)
     core.register_reaction('divide', divide_reaction)
-
-
-# Other functions
-
-def is_empty(value):
-    if isinstance(value, np.ndarray):
-        return False
-    elif value is None or value == {}:
-        return True
-    else:
-        return False
-
-def union_keys(schema, state):
-    keys = {}
-    for key in schema:
-        keys[key] = True
-    for key in state:
-        keys[key] = True
-
-    return keys
-    # return set(schema.keys()).union(state.keys())
-
-def tuple_from_type(tuple_type):
-    if isinstance(tuple_type, tuple):
-        return tuple_type
-
-    elif isinstance(tuple_type, list):
-        return tuple(tuple_type)
-
-    elif isinstance(tuple_type, dict):
-        tuple_list = [
-            tuple_type[f'_{parameter}']
-            for parameter in tuple_type['_type_parameters']]
-
-        return tuple(tuple_list)
-    else:
-        raise Exception(f'do not recognize this type as a tuple: {tuple_type}')
-
-def array_shape(core, schema):
-    if '_type_parameters' not in schema:
-        schema = core.access(schema)
-    parameters = schema.get('_type_parameters', [])
-
-    return tuple([
-        int(schema[f'_{parameter}'])
-        for parameter in schema['_type_parameters']])
-
-def lookup_dtype(data_name):
-    data_name = data_name or 'string'
-    dtype_name = DTYPE_MAP.get(data_name)
-    if dtype_name is None:
-        raise Exception(f'unknown data type for array: {data_name}')
-
-    return np.dtype(dtype_name)
-
-def read_datatype(data_schema):
-    return lookup_dtype(
-        data_schema['_type'])
-
-def read_shape(shape):
-    return tuple([
-        int(x)
-        for x in tuple_from_type(
-            shape)])
 
 
 # ===============================
